requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
textblob
cycler==0.10.0
docopt==0.6.2
Flask==2.0.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
gunicorn==20.0.2
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
numpy==1.20.3
pandas
Pillow
pipreqs
pyparsing
python-dateutil
pytz
requests
scikit-learn
scipy
six
sklearn
threadpoolctl
urllib3
Werkzeug
yarg
mlxtend
nltk
flask-login
requests
tensorflow
oauthlib
pyOpenSSL
keras
